Group ops add,subtract,... -> arithmetic.py, relational.py
