+++ temporian/core/operators/BUILD
@@ -24,8 +24,8 @@ py_library(
         ":set_index",
         ":since_last",
         ":unique_timestamps",
-        "//temporian/core/operators/arithmetic",
         "//temporian/core/operators/arithmetic_scalar",
+        "//temporian/core/operators/binary",
         "//temporian/core/operators/calendar:day_of_month",
         "//temporian/core/operators/calendar:day_of_week",
         "//temporian/core/operators/calendar:day_of_year",
