--- temporian/core/operators/arithmetic/BUILD
@@ -1,102 +1,0 @@
-package(
-    default_visibility = ["//visibility:public"],
-    licenses = ["notice"],
-)
-
-# Libraries
-# =========
-
-py_library(
-    name = "arithmetic",
-    srcs = ["__init__.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":add",
-        ":base",
-        ":divide",
-        ":equal",
-        ":floordiv",
-        ":multiply",
-        ":subtract",
-    ],
-)
-
-py_library(
-    name = "base",
-    srcs = ["base.py"],
-    srcs_version = "PY3",
-    deps = [
-        "//temporian/core/data:feature",
-        "//temporian/core/data:node",
-        "//temporian/core/operators:base",
-        "//temporian/proto:core_py_proto",
-    ],
-)
-
-py_library(
-    name = "add",
-    srcs = ["add.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":base",
-        "//temporian/core:operator_lib",
-        "//temporian/core/data:node",
-    ],
-)
-
-py_library(
-    name = "subtract",
-    srcs = ["subtract.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":base",
-        "//temporian/core:operator_lib",
-        "//temporian/core/data:node",
-    ],
-)
-
-py_library(
-    name = "multiply",
-    srcs = ["multiply.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":base",
-        "//temporian/core:operator_lib",
-        "//temporian/core/data:node",
-    ],
-)
-
-py_library(
-    name = "divide",
-    srcs = ["divide.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":base",
-        "//temporian/core:operator_lib",
-        "//temporian/core/data:node",
-    ],
-)
-
-py_library(
-    name = "equal",
-    srcs = ["equal.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":base",
-        "//temporian/core:operator_lib",
-        "//temporian/core/data:dtype",
-        "//temporian/core/data:feature",
-        "//temporian/core/data:node",
-    ],
-)
-
-py_library(
-    name = "floordiv",
-    srcs = ["floordiv.py"],
-    srcs_version = "PY3",
-    deps = [
-        ":base",
-        "//temporian/core:operator_lib",
-        "//temporian/core/data:node",
-    ],
-)
