+++ temporian/core/operators/binary/__init__.py
@@ -1,0 +1,18 @@
+# pylint: disable=unused-import
+
+from temporian.core.operators.binary.arithmetic import (
+    AddOperator,
+    SubtractOperator,
+    MultiplyOperator,
+    DivideOperator,
+    FloorDivOperator,
+    add,
+    subtract,
+    multiply,
+    divide,
+    floordiv,
+)
+
+from temporian.core.operators.binary.relational import EqualOperator, equal
+
+# pylint: enable=unused-import
