+++ temporian/core/operators/binary/relational.py
@@ -12,35 +12,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Subtract operator class and public API function definition."""
+"""Equal operator class and public API function definition."""
 
 from temporian.core import operator_lib
+from temporian.core.data.dtype import DType
 from temporian.core.data.node import Node
-from temporian.core.operators.arithmetic.base import BaseArithmeticOperator
+from temporian.core.data.feature import Feature
+from temporian.core.operators.binary.base import BaseBinaryOperator
 
 
-class SubtractOperator(BaseArithmeticOperator):
+class EqualOperator(BaseBinaryOperator):
     @classmethod
     @property
     def operator_def_key(cls) -> str:
-        return "SUBTRACTION"
+        return "EQUAL"
 
     @property
     def prefix(self) -> str:
-        return "sub"
+        return "equal"
+
+    # override parent dtype method
+    def output_feature_dtype(
+        self, feature_1: Feature, feature_2: Feature
+    ) -> DType:
+        return DType.BOOLEAN
 
 
-operator_lib.register_operator(SubtractOperator)
+operator_lib.register_operator(EqualOperator)
 
 
-def subtract(
+def equal(
     input_1: Node,
     input_2: Node,
 ) -> Node:
-    """Subtracts two nodes.
+    """Checks for equality between two nodes.
 
-    Each feature in `input_2` is subtracted from the feature in `input_1` in the
-    same position.
+    Each feature in `input_1` is compared element-wise to the feature in
+    `input_2` in the same position.
 
     `input_1` and `input_2` must have the same sampling and the same number of
     features.
@@ -50,9 +58,9 @@ def subtract(
         input_2: Second node.
 
     Returns:
-        Subtraction of `input_2`'s features from `input_1`'s.
+        Node containing the result of the comparison.
     """
-    return SubtractOperator(
+    return EqualOperator(
         input_1=input_1,
         input_2=input_2,
     ).outputs["output"]
