+++ temporian/implementation/numpy/operators/BUILD
@@ -25,13 +25,6 @@ py_library(
         ":set_index",
         ":since_last",
         ":unique_timestamps",
-        "//temporian/implementation/numpy/operators/arithmetic",
-        "//temporian/implementation/numpy/operators/arithmetic:add",
-        "//temporian/implementation/numpy/operators/arithmetic:divide",
-        "//temporian/implementation/numpy/operators/arithmetic:equal",
-        "//temporian/implementation/numpy/operators/arithmetic:floordiv",
-        "//temporian/implementation/numpy/operators/arithmetic:multiply",
-        "//temporian/implementation/numpy/operators/arithmetic:subtract",
         "//temporian/implementation/numpy/operators/arithmetic_scalar",
         "//temporian/implementation/numpy/operators/arithmetic_scalar:add",
         "//temporian/implementation/numpy/operators/arithmetic_scalar:divide",
@@ -41,6 +34,7 @@ py_library(
         "//temporian/implementation/numpy/operators/arithmetic_scalar:less",
         "//temporian/implementation/numpy/operators/arithmetic_scalar:multiply",
         "//temporian/implementation/numpy/operators/arithmetic_scalar:subtract",
+        "//temporian/implementation/numpy/operators/binary",
         "//temporian/implementation/numpy/operators/calendar:day_of_month",
         "//temporian/implementation/numpy/operators/calendar:day_of_week",
         "//temporian/implementation/numpy/operators/calendar:day_of_year",
