+++ temporian/implementation/numpy/operators/binary/__init__.py
@@ -1,0 +1,13 @@
+# pylint: disable=unused-import
+from temporian.implementation.numpy.operators.binary.arithmetic import (
+    AddNumpyImplementation,
+    SubtractNumpyImplementation,
+    MultiplyNumpyImplementation,
+    DivideNumpyImplementation,
+    FloorDivideNumpyImplementation,
+)
+from temporian.implementation.numpy.operators.binary.relational import (
+    EqualNumpyImplementation,
+)
+
+# pylint: enable=unused-import
