+++ temporian/implementation/numpy/operators/test/BUILD
@@ -12,22 +12,10 @@ py_test(
     deps = [
         "//temporian/core/data:node",
         "//temporian/core/data:sampling",
-        "//temporian/core/operators/arithmetic",
-        "//temporian/core/operators/arithmetic:add",
-        "//temporian/core/operators/arithmetic:divide",
-        "//temporian/core/operators/arithmetic:equal",
-        "//temporian/core/operators/arithmetic:floordiv",
-        "//temporian/core/operators/arithmetic:multiply",
-        "//temporian/core/operators/arithmetic:subtract",
+        "//temporian/core/operators/binary",
         "//temporian/implementation/numpy:evaluator",
         "//temporian/implementation/numpy/data:event_set",
-        "//temporian/implementation/numpy/operators/arithmetic",
-        "//temporian/implementation/numpy/operators/arithmetic:add",
-        "//temporian/implementation/numpy/operators/arithmetic:divide",
-        "//temporian/implementation/numpy/operators/arithmetic:equal",
-        "//temporian/implementation/numpy/operators/arithmetic:floordiv",
-        "//temporian/implementation/numpy/operators/arithmetic:multiply",
-        "//temporian/implementation/numpy/operators/arithmetic:subtract",
+        "//temporian/implementation/numpy/operators/binary",
     ],
 )
 
@@ -62,21 +50,9 @@ py_test(
         # absl/testing:absltest dep,
         "//temporian/core/data:node",
         "//temporian/core/data:sampling",
-        "//temporian/core/operators/arithmetic",
-        "//temporian/core/operators/arithmetic:add",
-        "//temporian/core/operators/arithmetic:subtract",
-        "//temporian/core/operators/arithmetic:multiply",
-        "//temporian/core/operators/arithmetic:divide",
-        "//temporian/core/operators/arithmetic:equal",
-        "//temporian/core/operators/arithmetic:floordiv",
+        "//temporian/core/operators/binary",
         "//temporian/implementation/numpy/data:event_set",
-        "//temporian/implementation/numpy/operators/arithmetic",
-        "//temporian/implementation/numpy/operators/arithmetic:add",
-        "//temporian/implementation/numpy/operators/arithmetic:subtract",
-        "//temporian/implementation/numpy/operators/arithmetic:multiply",
-        "//temporian/implementation/numpy/operators/arithmetic:divide",
-        "//temporian/implementation/numpy/operators/arithmetic:equal",
-        "//temporian/implementation/numpy/operators/arithmetic:floordiv",
+        "//temporian/implementation/numpy/operators/binary",
         "//temporian/implementation/numpy:evaluator",
     ],
 )
