+++ temporian/implementation/numpy/operators/test/arithmetic_multi_index_test.py
@@ -20,7 +20,7 @@ from temporian.core.data.dtype import DType
 from temporian.core.data.node import Node
 from temporian.core.data.node import Feature
 from temporian.core.data.sampling import Sampling
-from temporian.core.operators.arithmetic import (
+from temporian.core.operators.binary import (
     AddOperator,
     SubtractOperator,
     MultiplyOperator,
@@ -28,7 +28,7 @@ from temporian.core.operators.arithmetic import (
     FloorDivOperator,
 )
 from temporian.implementation.numpy.data.event_set import EventSet
-from temporian.implementation.numpy.operators.arithmetic import (
+from temporian.implementation.numpy.operators.binary import (
     AddNumpyImplementation,
     SubtractNumpyImplementation,
     MultiplyNumpyImplementation,
