+++ temporian/implementation/numpy/operators/test/arithmetic_test.py
@@ -19,7 +19,7 @@ from absl.testing import absltest
 from temporian.core.data.dtype import DType
 from temporian.core.data.node import Node, Feature
 from temporian.core.data.sampling import Sampling
-from temporian.core.operators.arithmetic import (
+from temporian.core.operators.binary import (
     AddOperator,
     SubtractOperator,
     MultiplyOperator,
@@ -27,7 +27,7 @@ from temporian.core.operators.arithmetic import (
     EqualOperator,
 )
 from temporian.implementation.numpy.data.event_set import EventSet
-from temporian.implementation.numpy.operators.arithmetic import (
+from temporian.implementation.numpy.operators.binary import (
     AddNumpyImplementation,
     SubtractNumpyImplementation,
     MultiplyNumpyImplementation,
